requirements.txt
Django==5.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.1
supabase==2.16.0
requests==2.32.4 